// website/src/lib/landing-assets.generated.ts
export const landingAssets = [
  {
    name: "3d-carousel",
    title: "3D Carousel",
    fileName: "3d-carousel-landing-asset.mp4",
    kind: "video",
    contentType: "video/mp4",
    bucketKey: "3d-carousel-landing-asset.mp4",
    bucketURL: "https://cdn.reacticx.com/3d-carousel-landing-asset.mp4?v=8e981377",
    size: 2662240,
    hash: "8e98137700238da1",
    width: 752,
    height: 844,
    aspect: 0.891,
  },
// ...
] as const satisfies readonly LandingAsset[];